Move list of hooks to config
